Change company name from 'Enter Green' to 'Enter Green building' in MobileApp component; adjust video container height and position of following section
This commit changes the company name from 'Enter Green' to 'Enter Green building' in the MobileApp component. Additionally, it adjusts the height and position of the video container and the following section. The height of the video container is increased from '278.5vh' to '280.5vh', and the margin top of the following section is increased from '0px' to '100px'. The border radius of the video container is also adjusted based on the value of the 'showVideo' variable.